Goodbye PPEx [4th October 2019]
Due to some issues with the license (do read [this](https://t.me/PaperplaneExtended/102)), I am now closing the Paperplane Extended project with this final update.

Here's what's changed in this update.

- 🔥 Added back the original Paperplane like syntax for notes, filters, welcomes, snips. (Now a valid BOTLOG_CHATID is only needed when we want to save media as data for these modules.)

[NOTE: This may require you to delete your current database and lose a lot of stuff, sorry for that.]

- Added some new env vars during initial deploy time.
`LOGSPAMMER` - If set to `True`, this makes the userbot throw the error.log file into the userbot log group instead of the current chat where the commad is triggered.
`WEATHER_DEFCITY` - Set a default city for the `.weather` module.
- Bumped up the telethon library to 1.10.6 to fix some issues.
- Async'd the google search module.
- Fixed the Magisk Release fetching module.
- Fixed the bug where the upload directory to Google Drive stuff acted strange.
- Added ability to add reason to bans, kicks, mutes, gmutes.
- Added ability to add custom rank for promotes.
- Merged the error_handler into the module wrapper.
- Added the RegexNinja module to prank people with sed bots :P
- Many optimisations under the hood. (as always.)

Now the whole codebase is archived, and will not be updated until further notice.
Thank you guys for your support, but it's time for this to go.

#PeaceOut :v: 
~~#StayExtended~~

+++ userbot/modules/purge.py
@@ -10,32 +10,34 @@ from asyncio import sleep
 from telethon.errors import rpcbaseerrors
 
 from userbot import BOTLOG, BOTLOG_CHATID, CMD_HELP
-from userbot.events import register, errors_handler
+from userbot.events import register
 
 
 @register(outgoing=True, pattern="^.purge$")
-@errors_handler
 async def fastpurger(purg):
     """ For .purge command, purge all messages starting from the reply. """
     chat = await purg.get_input_chat()
     msgs = []
+    itermsg = purg.client.iter_messages(chat, min_id=purg.reply_to_msg_id)
     count = 0
 
-    async for msg in purg.client.iter_messages(chat,
-                                               min_id=purg.reply_to_msg_id):
-        msgs.append(msg)
-        count = count + 1
-        msgs.append(purg.reply_to_msg_id)
-        if len(msgs) == 100:
-            await purg.client.delete_messages(chat, msgs)
-            msgs = []
+    if purg.reply_to_msg_id is not None:
+        async for msg in itermsg:
+            msgs.append(msg)
+            count = count + 1
+            msgs.append(purg.reply_to_msg_id)
+            if len(msgs) == 100:
+                await purg.client.delete_messages(chat, msgs)
+                msgs = []
+    else:
+        await purg.edit("`I need a mesasge to start purging from.`")
+        return
 
     if msgs:
         await purg.client.delete_messages(chat, msgs)
     done = await purg.client.send_message(
-        purg.chat_id,
-        "`Fast purge complete!\n`Purged " + str(count) + " messages.",
-    )
+        purg.chat_id, f"`Fast purge complete!`\
+        \nPurged {str(count)} messages")
 
     if BOTLOG:
         await purg.client.send_message(
@@ -46,7 +48,6 @@ async def fastpurger(purg):
 
 
 @register(outgoing=True, pattern="^.purgeme")
-@errors_handler
 async def purgeme(delme):
     """ For .purgeme, delete x count of your latest message."""
     message = delme.text
@@ -74,7 +75,6 @@ async def purgeme(delme):
 
 
 @register(outgoing=True, pattern="^.del$")
-@errors_handler
 async def delete_it(delme):
     """ For .del command, delete the replied message. """
     msg_src = await delme.get_reply_message()
@@ -92,7 +92,6 @@ async def delete_it(delme):
 
 
 @register(outgoing=True, pattern="^.edit")
-@errors_handler
 async def editer(edit):
     """ For .editme command, edit your last message. """
     message = edit.text
@@ -112,7 +111,6 @@ async def editer(edit):
 
 
 @register(outgoing=True, pattern="^.sd")
-@errors_handler
 async def selfdestruct(destroy):
     """ For .sd command, make seflf-destructable messages. """
     message = destroy.text
